Convert payment purchases history to use state from redux
